backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
groq==0.25.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.1
aiofiles==24.1.0
websockets==14.2
Pillow==11.1.0
aiosqlite==0.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.5